requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
